Assignment one of data types Modification

+++ src/Assignment_DataType.java
@@ -3,7 +3,7 @@ public class Assignment_DataType {
 
 	public static void main(String[] args) {
 		
-		double citytemp = 25.5;
+		float citytemp = 25.5f;
 		 boolean customer_order = true;
 		 String customer_phone= "123-456-7890";
 		 double customer_amountinbank=1000.50;
@@ -13,19 +13,18 @@ public class Assignment_DataType {
 		 String favourite_colour= "Blue";
 		 int current_year=2025;
 		 int followers_socialmedia =1000000;
-		  double movie_rating=7.5;
+		  float movie_rating=7.5f;
 		char blood_type='A';
 		String booktitle= "To Kill a Mockingbird";
 		short companyempy =500;
 		String country_name="United States";
 		String eye_color="Brown";
 		String birthplace= "New York City";
-        double city_dist=200.5;
+        float city_dist=200.5f;
         double locationlatitude =37.7749;
 		double location_longituda=-122.4194;
-		
 		//Time event variable
-		double event =2.30;
+		float event =2.30f;
 	}
 
 }
